+++ src/spec/views/result_item/TemporalMetadataView_spec.js
@@ -4,59 +4,53 @@ import _ from 'underscore';
 
 describe('NSIDC Temporal Metadata View', function () {
     describe('Content', function () {
-       it('keeps start and end when both exist', function () {
-           // one pair
-           // multiple pairs
-           //<dc:date>1973-01-01/2500-12-31</dc:date>
-           //<dc:date>2006-01-01/</dc:date>
-           let rangeList = [
-               ['1980-01-01', '2010-12-31'],
-               ['1990-01-01', '2020-12-31']
-           ];
-           let rangeArr = [];
+        function dateRangesToObject(rangeList) {
+            return _.map(rangeList, function (range) {
+                return {startDate: range[0], endDate: range[1]};
+            })
+        }
 
-           rangeArr = _.map(rangeList, function(range) {
-               return { startDate: range[0], endDate: range[1] };
-           })
-           expect(rangeArr.length).toEqual(2);
+        it('keeps start and end when both exist', function () {
+            let rangeList = [
+                ['1980-01-01', '2010-12-31'],
+                ['1990-01-01', '2020-12-31']
+            ];
 
-           let filtered = new TemporalMetadataView({}).filterRanges(rangeArr);
-           expect(filtered).toEqual(rangeArr);
-       });
+            let rangeArr = dateRangesToObject(rangeList);
+            expect(rangeArr.length).toEqual(2);
 
-       it('keeps only the start and shows present if end is empty', function () {
-           // one pair
-           // multiple pairs
-           let rangeList = [
-               ['1980-01-01', ''],
-               ['', '2020-12-31'],
-               ['', '']
-           ];
+            let filtered = new TemporalMetadataView({}).filterRanges(rangeArr);
+            expect(filtered).toEqual(rangeArr);
+        });
 
-           let rangeArr = [];
+        it('keeps only the start and shows present if end is empty', function () {
+            let rangeList = [
+                ['1970-01-01', '2000-12-31'],
+                ['1980-01-01', '']
+            ];
 
-           rangeArr = _.map(rangeList, function(range) {
-               return { startDate: range[0], endDate: range[1] };
-           })
-           expect(rangeArr.length).toEqual(3);
+            let rangeArr = dateRangesToObject(rangeList);
+            expect(rangeArr.length).toEqual(2);
 
-           let filtered = new TemporalMetadataView({}).filterRanges(rangeArr);
-           expect(filtered.length).toEqual(1);
-           expect(filtered[0]).toEqual(rangeArr[0]);
-           expect(filtered[0].startDate).toMatch('1980-01-01');
-           expect(filtered[0].endDate).toMatch('present');
-       });
+            let filtered = new TemporalMetadataView({}).filterRanges(rangeArr);
+            expect(filtered.length).toEqual(2);
+            expect(filtered[0].startDate).toMatch('1970-01-01');
+            expect(filtered[0].endDate).toMatch('2000-12-31');
+            expect(filtered[1].startDate).toMatch('1980-01-01');
+            expect(filtered[1].endDate).toMatch('present');
+        });
 
-       it('does not keep entries with missing start date', function () {
-           // empty string, undefined, not a date
-           // one pair
-           // multiple pairs
-           let rangeList = [
-               ['1980-01-01', ''],
-               ['', '2020-12-31'],
-               ['', '']
-           ];
-       });
+        it('does not keep entries with missing start date', function () {
+            let rangeList = [
+                ['', '2020-12-31'],
+                ['', '']
+            ];
+            let rangeArr = dateRangesToObject(rangeList);
+            expect(rangeArr.length).toEqual(2);
+
+            let filtered = new TemporalMetadataView({}).filterRanges(rangeArr);
+            expect(filtered.length).toEqual(0);
+        });
     });
 
     describe('Rendering', function () {